+++ cmd/jobsShow.go
@@ -118,8 +118,14 @@ func PrintJobTransfers(listTransfersResponse common.ListJobTransfersResponse) {
 		var sb strings.Builder
 		sb.WriteString("----------- Transfers for JobId " + listTransfersResponse.JobID.String() + " -----------\n")
 		for index := 0; index < len(listTransfersResponse.Details); index++ {
-			sb.WriteString("transfer--> source: " + listTransfersResponse.Details[index].Src + " destination: " +
-				listTransfersResponse.Details[index].Dst + " status " + listTransfersResponse.Details[index].TransferStatus.String() + "\n")
+			folderChar := ""
+			if listTransfersResponse.Details[index].IsFolderProperties {
+				folderChar = "/"
+				// TODO: reviewers, is it OK that this is hard coded to fwd slash (even for windows local sources?)
+				// TODO: reviewers, would it be better to output something like "type: folder"?  I did't because it seemed verbose.
+			}
+			sb.WriteString("transfer--> source: " + listTransfersResponse.Details[index].Src + folderChar + " destination: " +
+				listTransfersResponse.Details[index].Dst + folderChar + " status " + listTransfersResponse.Details[index].TransferStatus.String() + "\n")
 		}
 
 		return sb.String()